+++ android/common/src/main/java/com/marianhello/bgloc/data/sqlite/SQLiteConfigurationContract.java
@@ -45,6 +45,8 @@ public final class SQLiteConfigurationContract {
         public static final String COLUMN_NAME_HEADERS = "http_headers";
         public static final String COLUMN_NAME_MAX_LOCATIONS = "max_locations";
         public static final String COLUMN_NAME_TEMPLATE = "template";
+        public static final String COLUMN_NAME_SHOW_TIME = "show_time";
+        public static final String COLUMN_NAME_SHOW_DISTANCE = "show_distance";
 
         public static final String SQL_CREATE_CONFIG_TABLE =
                 "CREATE TABLE " + ConfigurationEntry.TABLE_NAME + " (" +
@@ -77,7 +79,9 @@ public final class SQLiteConfigurationContract {
                         ConfigurationEntry.COLUMN_NAME_SYNC_ENABLED + INTEGER_TYPE + COMMA_SEP +
                         ConfigurationEntry.COLUMN_NAME_HEADERS + TEXT_TYPE + COMMA_SEP +
                         ConfigurationEntry.COLUMN_NAME_MAX_LOCATIONS + INTEGER_TYPE + COMMA_SEP +
-                        ConfigurationEntry.COLUMN_NAME_TEMPLATE + TEXT_TYPE +
+                        ConfigurationEntry.COLUMN_NAME_TEMPLATE + TEXT_TYPE + COMMA_SEP +
+                        ConfigurationEntry.COLUMN_NAME_SHOW_TIME + INTEGER_TYPE + COMMA_SEP +
+                        ConfigurationEntry.COLUMN_NAME_SHOW_DISTANCE + INTEGER_TYPE +
                         " )";
 
         public static final String SQL_DROP_CONFIG_TABLE =